rename test class, restructured base test class for ToDotemsController
