+++ ToDoList/tests/ToDoList.Test/ActionResultExtensions.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 
 public static class ActionResultExtensions
 {
-    public static T? GetValue<T>(this ActionResult<T> result) => result.Result is null
-        ? result.Value
-        : (T?)(result.Result as ObjectResult)?.Value;
+    public static T? GetValue<T>(this ActionResult<T> response) => response.Result is null
+        ? response.Value
+        : (T?)(response.Result as ObjectResult)?.Value;
 }
