--- ToDoList/tests/ToDoList.Test/GetTests.cs
@@ -1,43 +1,0 @@
-namespace ToDoList.Test;
-
-using ToDoList.Domain.Models;
-using ToDoList.WebApi;
-
-public class GetTests
-{
-    [Fact]
-    public void Get_AllItems_ReturnsAllItems()
-    {
-        // Arrange
-        var todoItem1 = new ToDoItem
-        {
-            ToDoItemId = 1,
-            Name = "Jmeno1",
-            Description = "Popis1",
-            IsCompleted = false
-        };
-        var todoItem2 = new ToDoItem
-        {
-            ToDoItemId = 2,
-            Name = "Jmeno2",
-            Description = "Popis2",
-            IsCompleted = true
-        };
-        var controller = new ToDoItemsController();
-        controller.AddItemToStorage(todoItem1);
-        controller.AddItemToStorage(todoItem2);
-
-        // Act
-        var result = controller.Read();
-        var value = result.GetValue();
-
-        // Assert
-        Assert.NotNull(value);
-
-        var firstToDo = value.First();
-        Assert.Equal(todoItem1.ToDoItemId, firstToDo.Id);
-        Assert.Equal(todoItem1.Name, firstToDo.Name);
-        Assert.Equal(todoItem1.Description, firstToDo.Description);
-        Assert.Equal(todoItem1.IsCompleted, firstToDo.IsCompleted);
-    }
-}
